+++ service/user/data/request/doc/doc_group.go
@@ -3,7 +3,7 @@ package doc
 // CreateDocGroupRequest 创建文档分组
 type CreateDocGroupRequest struct {
 	Title    string `json:"title" binding:"required,min=1,max=250"` // 标题
-	PId      int    `json:"pId" binding:""`                         // 父级
+	PId      int64  `json:"pId" binding:""`                         // 父级
 	Icon     string `json:"icon" binding:""`                        // 图标
 	Priority int    `json:"priority" binding:""`
 }
